+++ example/lib/place_circle.dart
@@ -145,7 +145,8 @@ class PlaceCircleBodyState extends State<PlaceCircleBody> {
       // default value
       current = 0;
     }
-    _updateSelectedCircle(CircleOptions(circleStrokeWidth: current == 0 ? 5.0 : 0));
+    _updateSelectedCircle(
+        CircleOptions(circleStrokeWidth: current == 0 ? 5.0 : 0));
   }
 
   Future<void> _changeCircleStrokeColor() async {
@@ -156,7 +157,8 @@ class PlaceCircleBodyState extends State<PlaceCircleBody> {
     }
 
     _updateSelectedCircle(
-      CircleOptions(circleStrokeColor: current == "#FFFFFF" ? "#FF0000" : "#FFFFFF"),
+      CircleOptions(
+          circleStrokeColor: current == "#FFFFFF" ? "#FF0000" : "#FFFFFF"),
     );
   }
 
@@ -191,8 +193,7 @@ class PlaceCircleBodyState extends State<PlaceCircleBody> {
     }
 
     _updateSelectedCircle(
-      CircleOptions(
-          circleColor: "#FFFF00"),
+      CircleOptions(circleColor: "#FFFF00"),
     );
   }
 
@@ -250,8 +251,9 @@ class PlaceCircleBodyState extends State<PlaceCircleBody> {
                       children: <Widget>[
                         TextButton(
                           child: const Text('change circle-opacity'),
-                          onPressed:
-                              (_selectedCircle == null) ? null : _changeCircleOpacity,
+                          onPressed: (_selectedCircle == null)
+                              ? null
+                              : _changeCircleOpacity,
                         ),
                         TextButton(
                           child: const Text('change circle-radius'),
@@ -261,18 +263,21 @@ class PlaceCircleBodyState extends State<PlaceCircleBody> {
                         ),
                         TextButton(
                           child: const Text('change circle-color'),
-                          onPressed:
-                          (_selectedCircle == null) ? null : _changeCircleColor,
+                          onPressed: (_selectedCircle == null)
+                              ? null
+                              : _changeCircleColor,
                         ),
                         TextButton(
                           child: const Text('change circle-blur'),
-                          onPressed:
-                          (_selectedCircle == null) ? null : _changeCircleBlur,
+                          onPressed: (_selectedCircle == null)
+                              ? null
+                              : _changeCircleBlur,
                         ),
                         TextButton(
                           child: const Text('change circle-stroke-width'),
-                          onPressed:
-                              (_selectedCircle == null) ? null : _changeCircleStrokeWidth,
+                          onPressed: (_selectedCircle == null)
+                              ? null
+                              : _changeCircleStrokeWidth,
                         ),
                         TextButton(
                           child: const Text('change circle-stroke-color'),
@@ -300,9 +305,8 @@ class PlaceCircleBodyState extends State<PlaceCircleBody> {
                         ),
                         TextButton(
                           child: const Text('get current LatLng'),
-                          onPressed: (_selectedCircle == null)
-                              ? null
-                              : _getLatLng,
+                          onPressed:
+                              (_selectedCircle == null) ? null : _getLatLng,
                         ),
                       ],
                     ),