+++ ios/Classes/MapboxMapController.swift
@@ -23,6 +23,7 @@ class MapboxMapController: NSObject, FlutterPlatformView, MGLMapViewDelegate, Ma
     private var fillAnnotationController: MGLPolygonAnnotationController?
 
     private var annotationOrder = [String]()
+    private var annotationConsumeTapEvents = [String]()
 
     func view() -> UIView {
         return mapView
@@ -63,9 +64,17 @@ class MapboxMapController: NSObject, FlutterPlatformView, MGLMapViewDelegate, Ma
             if let annotationOrderArg = args["annotationOrder"] as? [String] {
                 annotationOrder = annotationOrderArg
             }
+            if let annotationConsumeTapEventsArg = args["annotationConsumeTapEvents"] as? [String] {
+                annotationConsumeTapEvents = annotationConsumeTapEventsArg
+            }
         }
     }
-    
+    func removeAllForController(controller: MGLAnnotationController, ids: [String]){
+        let idSet = Set(ids)
+        let annotations = controller.styleAnnotations()
+        controller.removeStyleAnnotations(annotations.filter { idSet.contains($0.identifier) })
+    }
+
     func onMethodCall(methodCall: FlutterMethodCall, result: @escaping FlutterResult) {
         switch(methodCall.method) {
         case "map#waitForMap":
@@ -176,6 +185,25 @@ class MapboxMapController: NSObject, FlutterPlatformView, MGLMapViewDelegate, Ma
             reply["x"] = returnVal.x as NSObject
             reply["y"] = returnVal.y as NSObject
             result(reply)
+        case "map#toScreenLocationBatch":
+            guard let arguments = methodCall.arguments as? [String: Any] else { return }
+            guard let data = arguments["coordinates"] as? FlutterStandardTypedData else { return }
+            let latLngs = data.data.withUnsafeBytes {
+                Array(
+                    UnsafeBufferPointer(
+                        start: $0.baseAddress!.assumingMemoryBound(to: Double.self),
+                        count:Int(data.elementCount))
+                )
+            }
+            var reply: [Double] = Array(repeating: 0.0, count: latLngs.count)
+            for i in stride(from: 0, to: latLngs.count, by: 2) {
+                let coordinate = CLLocationCoordinate2DMake(latLngs[i], latLngs[i + 1])
+                let returnVal = mapView.convert(coordinate, toPointTo: mapView)
+                reply[i] = Double(returnVal.x)
+                reply[i + 1] = Double(returnVal.y)
+            }
+            result(FlutterStandardTypedData(
+                    float64: Data(bytes: &reply, count: reply.count * 8) ))
         case "map#getMetersPerPixelAtLatitude":
              guard let arguments = methodCall.arguments as? [String: Any] else { return }
              var reply = [String: NSObject]()
@@ -225,6 +253,7 @@ class MapboxMapController: NSObject, FlutterPlatformView, MGLMapViewDelegate, Ma
                 }
                 if !symbols.isEmpty {
                     symbolAnnotationController.addStyleAnnotations(symbols)
+                    symbolAnnotationController.annotationsInteractionEnabled = annotationConsumeTapEvents.contains("AnnotationType.symbol")
                 }
 
                 result(symbols.map { $0.identifier })
@@ -252,16 +281,11 @@ class MapboxMapController: NSObject, FlutterPlatformView, MGLMapViewDelegate, Ma
         case "symbols#removeAll":
             guard let symbolAnnotationController = symbolAnnotationController else { return }
             guard let arguments = methodCall.arguments as? [String: Any] else { return }
-            guard let symbolIds = arguments["symbols"] as? [String] else { return }
-            var symbols: [MGLSymbolStyleAnnotation] = [];
+            guard let symbolIds = arguments["ids"] as? [String] else { return }
 
-            for symbol in symbolAnnotationController.styleAnnotations(){
-                if symbolIds.contains(symbol.identifier) {
-                    symbols.append(symbol as! MGLSymbolStyleAnnotation)
-                }
-            }
-            symbolAnnotationController.removeStyleAnnotations(symbols)
+            removeAllForController(controller:symbolAnnotationController, ids:symbolIds)
             result(nil)
+
         case "symbol#getGeometry":
             guard let symbolAnnotationController = symbolAnnotationController else { return }
             guard let arguments = methodCall.arguments as? [String: Any] else { return }
@@ -312,10 +336,39 @@ class MapboxMapController: NSObject, FlutterPlatformView, MGLMapViewDelegate, Ma
                 let circle = MGLCircleStyleAnnotation(center: coordinate)
                 Convert.interpretCircleOptions(options: arguments["options"], delegate: circle)
                 circleAnnotationController.addStyleAnnotation(circle)
+                circleAnnotationController.annotationsInteractionEnabled = annotationConsumeTapEvents.contains("AnnotationType.circle")
                 result(circle.identifier)
             } else {
                 result(nil)
             }
+
+        case "circle#addAll":
+            guard let circleAnnotationController = circleAnnotationController else { return }
+            guard let arguments = methodCall.arguments as? [String: Any] else { return }
+            // Parse geometry
+            var identifier: String? = nil
+            if let allOptions = arguments["options"] as? [[String: Any]]{
+                var circles: [MGLCircleStyleAnnotation] = [];
+
+                for options in allOptions {
+                    if let geometry = options["geometry"] as? [Double] {
+                        guard geometry.count > 0 else { break }
+
+                        let coordinate = CLLocationCoordinate2DMake(geometry[0], geometry[1])
+                        let circle = MGLCircleStyleAnnotation(center: coordinate)
+                        Convert.interpretCircleOptions(options: options, delegate: circle)
+                        circles.append(circle)
+                    }
+                }
+                if !circles.isEmpty {
+                    circleAnnotationController.addStyleAnnotations(circles)
+                }
+                result(circles.map { $0.identifier })
+            }
+            else {
+                result(nil)
+            }
+
         case "circle#update":
             guard let circleAnnotationController = circleAnnotationController else { return }
             guard let arguments = methodCall.arguments as? [String: Any] else { return }
@@ -341,6 +394,16 @@ class MapboxMapController: NSObject, FlutterPlatformView, MGLMapViewDelegate, Ma
                 }
             }
             result(nil)
+
+        case "circle#removeAll":
+            guard let circleAnnotationController = circleAnnotationController else { return }
+            guard let arguments = methodCall.arguments as? [String: Any] else { return }
+            guard let ids = arguments["ids"] as? [String] else { return }
+
+            removeAllForController(controller:circleAnnotationController, ids:ids)
+            result(nil)
+
+
         case "line#add":
             guard let lineAnnotationController = lineAnnotationController else { return }
             guard let arguments = methodCall.arguments as? [String: Any] else { return }
@@ -355,10 +418,43 @@ class MapboxMapController: NSObject, FlutterPlatformView, MGLMapViewDelegate, Ma
                 let line = MGLLineStyleAnnotation(coordinates: lineCoordinates, count: UInt(lineCoordinates.count))
                 Convert.interpretLineOptions(options: arguments["options"], delegate: line)
                 lineAnnotationController.addStyleAnnotation(line)
+                lineAnnotationController.annotationsInteractionEnabled = annotationConsumeTapEvents.contains("AnnotationType.line")
                 result(line.identifier)
             } else {
                 result(nil)
             }
+
+        case "line#addAll":
+            guard let lineAnnotationController = lineAnnotationController else { return }
+            guard let arguments = methodCall.arguments as? [String: Any] else { return }
+            // Parse geometry
+            var identifier: String? = nil
+            if let allOptions = arguments["options"] as? [[String: Any]]{
+                var lines: [MGLLineStyleAnnotation] = [];
+
+                for options in allOptions {
+                    if let geometry = options["geometry"] as? [[Double]] {
+                        guard geometry.count > 0 else { break }
+                        // Convert geometry to coordinate and create a line.
+                        var lineCoordinates: [CLLocationCoordinate2D] = []
+                        for coordinate in geometry {
+                            lineCoordinates.append(CLLocationCoordinate2DMake(coordinate[0], coordinate[1]))
+                        }
+                        let line = MGLLineStyleAnnotation(coordinates: lineCoordinates, count: UInt(lineCoordinates.count))
+                        Convert.interpretLineOptions(options: options, delegate: line)
+                        lines.append(line)
+                    }
+                }
+                if !lines.isEmpty {
+                    lineAnnotationController.addStyleAnnotations(lines)
+                }
+                result(lines.map { $0.identifier })
+            }
+            else {
+                result(nil)
+            }
+
+
         case "line#update":
             guard let lineAnnotationController = lineAnnotationController else { return }
             guard let arguments = methodCall.arguments as? [String: Any] else { return }
@@ -384,6 +480,15 @@ class MapboxMapController: NSObject, FlutterPlatformView, MGLMapViewDelegate, Ma
                 }
             }
             result(nil)
+
+        case "line#removeAll":
+            guard let lineAnnotationController = lineAnnotationController else { return }
+            guard let arguments = methodCall.arguments as? [String: Any] else { return }
+            guard let ids = arguments["ids"] as? [String] else { return }
+
+            removeAllForController(controller:lineAnnotationController, ids:ids)
+            result(nil)
+
         case "line#getGeometry":
             guard let lineAnnotationController = lineAnnotationController else { return }
             guard let arguments = methodCall.arguments as? [String: Any] else { return }
@@ -417,9 +522,43 @@ class MapboxMapController: NSObject, FlutterPlatformView, MGLMapViewDelegate, Ma
                 let fill = MGLPolygonStyleAnnotation(coordinates: fillCoordinates, count: UInt(fillCoordinates.count), interiorPolygons: polygons)
                 Convert.interpretFillOptions(options: arguments["options"], delegate: fill)
                 fillAnnotationController.addStyleAnnotation(fill)
+                fillAnnotationController.annotationsInteractionEnabled = annotationConsumeTapEvents.contains("AnnotationType.fill")
                 identifier = fill.identifier
             }
+
             result(identifier)
+
+        case "fill#addAll":
+            guard let fillAnnotationController = fillAnnotationController else { return }
+            guard let arguments = methodCall.arguments as? [String: Any] else { return }
+            // Parse geometry
+            var identifier: String? = nil
+            if let allOptions = arguments["options"] as? [[String: Any]]{
+                var fills: [MGLPolygonStyleAnnotation] = [];
+
+                for options in allOptions{
+                    if let geometry = options["geometry"] as? [[[Double]]] {
+                        guard geometry.count > 0 else { break }
+                        // Convert geometry to coordinate and interior polygonc.
+                        var fillCoordinates: [CLLocationCoordinate2D] = []
+                        for coordinate in geometry[0] {
+                            fillCoordinates.append(CLLocationCoordinate2DMake(coordinate[0], coordinate[1]))
+                        }
+                        let polygons = Convert.toPolygons(geometry: geometry.tail)
+                        let fill = MGLPolygonStyleAnnotation(coordinates: fillCoordinates, count: UInt(fillCoordinates.count), interiorPolygons: polygons)
+                        Convert.interpretFillOptions(options: options, delegate: fill)
+                        fills.append(fill)
+                    }
+                }
+                if !fills.isEmpty {
+                    fillAnnotationController.addStyleAnnotations(fills)
+                }
+                result(fills.map { $0.identifier })
+            }
+            else {
+                result(nil)
+            }
+
         case "fill#update":
             guard let fillAnnotationController = fillAnnotationController else { return }
             guard let arguments = methodCall.arguments as? [String: Any] else { return }
@@ -432,6 +571,7 @@ class MapboxMapController: NSObject, FlutterPlatformView, MGLMapViewDelegate, Ma
                     break;
                 }
             }
+
             result(nil)
         case "fill#remove":
             guard let fillAnnotationController = fillAnnotationController else { return }
@@ -445,6 +585,15 @@ class MapboxMapController: NSObject, FlutterPlatformView, MGLMapViewDelegate, Ma
                 }
             }
             result(nil)
+
+        case "fill#removeAll":
+            guard let fillAnnotationController = fillAnnotationController else { return }
+            guard let arguments = methodCall.arguments as? [String: Any] else { return }
+            guard let ids = arguments["ids"] as? [String] else { return }
+
+            removeAllForController(controller:fillAnnotationController, ids:ids)
+            result(nil)
+
         case "style#addImage":
             guard let arguments = methodCall.arguments as? [String: Any] else { return }
             guard let name = arguments["name"] as? String else { return }
@@ -459,6 +608,8 @@ class MapboxMapController: NSObject, FlutterPlatformView, MGLMapViewDelegate, Ma
                 self.mapView.style?.setImage(image, forName: name)
             }
             result(nil)
+
+
         case "style#addImageSource":
             guard let arguments = methodCall.arguments as? [String: Any] else { return }
             guard let imageSourceId = arguments["imageSourceId"] as? String else { return }